+++ packages/git/src/application/adapter/GitAdapter.ts
@@ -40,26 +40,26 @@ import { IGitDelayedJobs } from '../../domain/jobs/IGitDelayedJobs';
 import { FetchFileContentJobFactory } from '../../infra/jobs/FetchFileContentJobFactory';
 import { GithubAppMode } from '../../infra/repositories/github/auth/GithubTokenResolverFactory';
 import { GitServices } from '../GitServices';
-import { AddGitProviderUseCase } from '../useCases/addGitProvider/addGitProvider.usecase';
-import { AddGitRepoUseCase } from '../useCases/addGitRepo/addGitRepo.usecase';
-import { CheckBranchExistsUseCase } from '../useCases/checkBranchExists/checkBranchExists.usecase';
-import { CheckDirectoryExistenceUseCase } from '../useCases/checkDirectoryExistence/checkDirectoryExistence.usecase';
-import { CheckProviderAuthUseCase } from '../useCases/checkProviderAuth/checkProviderAuth.usecase';
-import { CommitToGit } from '../useCases/commitToGit/commitToGit.usecase';
-import { DeleteGitProviderUseCase } from '../useCases/deleteGitProvider/deleteGitProvider.usecase';
-import { DeleteGitRepoUseCase } from '../useCases/deleteGitRepo/deleteGitRepo.usecase';
-import { FindGitRepoByOwnerAndRepoUseCase } from '../useCases/findGitRepoByOwnerAndRepo/findGitRepoByOwnerAndRepo.usecase';
-import { FindGitRepoByOwnerRepoAndBranchInOrganizationUseCase } from '../useCases/findGitRepoByOwnerRepoAndBranchInOrganization/findGitRepoByOwnerRepoAndBranchInOrganization.usecase';
-import { GetAvailableRemoteDirectoriesUseCase } from '../useCases/getAvailableRemoteDirectories/getAvailableRemoteDirectories.usecase';
-import { GetFileFromRepo } from '../useCases/getFileFromRepo/getFileFromRepo.usecase';
-import { GetOrganizationRepositoriesUseCase } from '../useCases/getOrganizationRepositories/getOrganizationRepositories.usecase';
-import { GetRepositoryByIdUseCase } from '../useCases/getRepositoryById/getRepositoryById.usecase';
-import { HandleWebHook } from '../useCases/handleWebHook/handleWebHook.usecase';
-import { HandleWebHookWithoutContent } from '../useCases/handleWebHookWithoutContent/handleWebHookWithoutContent.usecase';
-import { ListAvailableReposUseCase } from '../useCases/listAvailableRepos/listAvailableRepos.usecase';
-import { ListProvidersUseCase } from '../useCases/listProviders/listProviders.usecase';
-import { ListReposUseCase } from '../useCases/listRepos/listRepos.usecase';
-import { UpdateGitProviderUseCase } from '../useCases/updateGitProvider/updateGitProvider.usecase';
+import { AddGitProviderUseCase } from '../useCases/addGitProvider/AddGitProviderUseCase';
+import { AddGitRepoUseCase } from '../useCases/addGitRepo/AddGitRepoUseCase';
+import { CheckBranchExistsUseCase } from '../useCases/checkBranchExists/CheckBranchExistsUseCase';
+import { CheckDirectoryExistenceUseCase } from '../useCases/checkDirectoryExistence/CheckDirectoryExistenceUseCase';
+import { CheckProviderAuthUseCase } from '../useCases/checkProviderAuth/CheckProviderAuthUseCase';
+import { CommitToGit } from '../useCases/commitToGit/CommitToGit';
+import { DeleteGitProviderUseCase } from '../useCases/deleteGitProvider/DeleteGitProviderUseCase';
+import { DeleteGitRepoUseCase } from '../useCases/deleteGitRepo/DeleteGitRepoUseCase';
+import { FindGitRepoByOwnerAndRepoUseCase } from '../useCases/findGitRepoByOwnerAndRepo/FindGitRepoByOwnerAndRepoUseCase';
+import { FindGitRepoByOwnerRepoAndBranchInOrganizationUseCase } from '../useCases/findGitRepoByOwnerRepoAndBranchInOrganization/FindGitRepoByOwnerRepoAndBranchInOrganizationUseCase';
+import { GetAvailableRemoteDirectoriesUseCase } from '../useCases/getAvailableRemoteDirectories/GetAvailableRemoteDirectoriesUseCase';
+import { GetFileFromRepo } from '../useCases/getFileFromRepo/GetFileFromRepo';
+import { GetOrganizationRepositoriesUseCase } from '../useCases/getOrganizationRepositories/GetOrganizationRepositoriesUseCase';
+import { GetRepositoryByIdUseCase } from '../useCases/getRepositoryById/GetRepositoryByIdUseCase';
+import { HandleWebHook } from '../useCases/handleWebHook/HandleWebHook';
+import { HandleWebHookWithoutContent } from '../useCases/handleWebHookWithoutContent/HandleWebHookWithoutContent';
+import { ListAvailableReposUseCase } from '../useCases/listAvailableRepos/ListAvailableReposUseCase';
+import { ListProvidersUseCase } from '../useCases/listProviders/ListProvidersUseCase';
+import { ListReposUseCase } from '../useCases/listRepos/ListReposUseCase';
+import { UpdateGitProviderUseCase } from '../useCases/updateGitProvider/UpdateGitProviderUseCase';
 
 const origin = 'GitAdapter';
