+++ packages/git/src/application/useCases/index.ts
@@ -2,22 +2,22 @@
 export {
   AddGitProviderUseCase,
   AddGitProviderCommand,
-} from './addGitProvider/addGitProvider.usecase';
-export { AddGitRepoUseCase } from './addGitRepo/addGitRepo.usecase';
-export { CheckDirectoryExistenceUseCase } from './checkDirectoryExistence/checkDirectoryExistence.usecase';
-export { DeleteGitProviderUseCase } from './deleteGitProvider/deleteGitProvider.usecase';
-export { DeleteGitRepoUseCase } from './deleteGitRepo/deleteGitRepo.usecase';
-export { ListAvailableReposUseCase } from './listAvailableRepos/listAvailableRepos.usecase';
-export { CheckBranchExistsUseCase } from './checkBranchExists/checkBranchExists.usecase';
-export { CommitToGit } from './commitToGit/commitToGit.usecase';
-export { HandleWebHook } from './handleWebHook/handleWebHook.usecase';
-export { HandleWebHookWithoutContent } from './handleWebHookWithoutContent/handleWebHookWithoutContent.usecase';
-export { GetFileFromRepo } from './getFileFromRepo/getFileFromRepo.usecase';
-export { FindGitRepoByOwnerAndRepoUseCase } from './findGitRepoByOwnerAndRepo/findGitRepoByOwnerAndRepo.usecase';
-export { ListReposUseCase } from './listRepos/listRepos.usecase';
-export { ListProvidersUseCase } from './listProviders/listProviders.usecase';
-export { GetOrganizationRepositoriesUseCase } from './getOrganizationRepositories/getOrganizationRepositories.usecase';
-export { GetRepositoryByIdUseCase } from './getRepositoryById/getRepositoryById.usecase';
-export { UpdateGitProviderUseCase } from './updateGitProvider/updateGitProvider.usecase';
-export { GetAvailableRemoteDirectoriesUseCase } from './getAvailableRemoteDirectories/getAvailableRemoteDirectories.usecase';
-export { FindGitRepoByOwnerRepoAndBranchInOrganizationUseCase } from './findGitRepoByOwnerRepoAndBranchInOrganization/findGitRepoByOwnerRepoAndBranchInOrganization.usecase';
+} from './addGitProvider/AddGitProviderUseCase';
+export { AddGitRepoUseCase } from './addGitRepo/AddGitRepoUseCase';
+export { CheckDirectoryExistenceUseCase } from './checkDirectoryExistence/CheckDirectoryExistenceUseCase';
+export { DeleteGitProviderUseCase } from './deleteGitProvider/DeleteGitProviderUseCase';
+export { DeleteGitRepoUseCase } from './deleteGitRepo/DeleteGitRepoUseCase';
+export { ListAvailableReposUseCase } from './listAvailableRepos/ListAvailableReposUseCase';
+export { CheckBranchExistsUseCase } from './checkBranchExists/CheckBranchExistsUseCase';
+export { CommitToGit } from './commitToGit/CommitToGit';
+export { HandleWebHook } from './handleWebHook/HandleWebHook';
+export { HandleWebHookWithoutContent } from './handleWebHookWithoutContent/HandleWebHookWithoutContent';
+export { GetFileFromRepo } from './getFileFromRepo/GetFileFromRepo';
+export { FindGitRepoByOwnerAndRepoUseCase } from './findGitRepoByOwnerAndRepo/FindGitRepoByOwnerAndRepoUseCase';
+export { ListReposUseCase } from './listRepos/ListReposUseCase';
+export { ListProvidersUseCase } from './listProviders/ListProvidersUseCase';
+export { GetOrganizationRepositoriesUseCase } from './getOrganizationRepositories/GetOrganizationRepositoriesUseCase';
+export { GetRepositoryByIdUseCase } from './getRepositoryById/GetRepositoryByIdUseCase';
+export { UpdateGitProviderUseCase } from './updateGitProvider/UpdateGitProviderUseCase';
+export { GetAvailableRemoteDirectoriesUseCase } from './getAvailableRemoteDirectories/GetAvailableRemoteDirectoriesUseCase';
+export { FindGitRepoByOwnerRepoAndBranchInOrganizationUseCase } from './findGitRepoByOwnerRepoAndBranchInOrganization/FindGitRepoByOwnerRepoAndBranchInOrganizationUseCase';
